Remove preprocessing dependencies
All LS-features that depended on preprocessing depend on sketch instead
